remove hardcoded dtype (#1689)
this PRs removes legacy hardcoded int32 types

Authors:
  - https://github.com/Iroy30

Approvers:
  - Brad Rees (https://github.com/BradReesWork)
  - Rick Ratzel (https://github.com/rlratzel)

URL: https://github.com/rapidsai/cugraph/pull/1689

+++ python/cugraph/dask/traversal/bfs.py
@@ -110,8 +110,8 @@ def bfs(graph,
                     compute()
             start = start.iloc[0]
         else:
-            start = graph.lookup_internal_vertex_id(cudf.Series([start],
-                                                    dtype='int32')).compute()
+            start = graph.lookup_internal_vertex_id(cudf.Series([start])
+                                                    ).compute()
             start = start.iloc[0]
 
     result = [client.submit(
